[Refactor/#7] 기본 세팅
