+++ app/src/main/java/org/sopt/at/navigation/HomeNavGraph.kt
@@ -3,11 +3,11 @@ package org.sopt.at.navigation
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.composable
-import org.sopt.at.history.HistoryScreen
-import org.sopt.at.home.HomeScreen
-import org.sopt.at.live.LiveScreen
-import org.sopt.at.search.SearchScreen
-import org.sopt.at.shorts.ShortsScreen
+import org.sopt.at.presentation.history.HistoryScreen
+import org.sopt.at.presentation.home.HomeScreen
+import org.sopt.at.presentation.live.LiveScreen
+import org.sopt.at.presentation.search.SearchScreen
+import org.sopt.at.presentation.shorts.ShortsScreen
 
 fun NavGraphBuilder.homeNavGraph(navController: NavHostController) {
     composable("HomeScreen") { HomeScreen(navController) }
