+++ app/src/main/java/org/sopt/at/navigation/MyNavGraph.kt
@@ -3,7 +3,7 @@ package org.sopt.at.navigation
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.composable
-import org.sopt.at.my.MyScreen
+import org.sopt.at.presentation.my.MyScreen
 
 fun NavGraphBuilder.myNavGraph(navController: NavHostController) {
     composable("MyScreen/{userId}") { backStackEntry ->
