+++ app/src/main/java/org/sopt/at/navigation/SignUpNavGraph.kt
@@ -6,23 +6,23 @@ import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.composable
-import org.sopt.at.signup.IdInputScreen
-import org.sopt.at.signup.PasswordInputScreen
-import org.sopt.at.signup.SignUpViewModel
+import org.sopt.at.presentation.signup.IdInputScreen
+import org.sopt.at.presentation.signup.PasswordInputScreen
+import org.sopt.at.presentation.signup.SignUpViewModel
 
 fun NavGraphBuilder.signupNavGraph(navController: NavHostController) {
     composable("IdInputScreen") { backStackEntry ->
         val parentEntry = remember (backStackEntry){
             navController.getBackStackEntry("IdInputScreen")
         }
-        val viewModel:SignUpViewModel = viewModel(parentEntry)
+        val viewModel: SignUpViewModel = viewModel(parentEntry)
         IdInputScreen(navController,viewModel)
     }
     composable("PasswordInputScreen"){ backStackEntry ->
         val parentEntry = remember(backStackEntry){
             navController.getBackStackEntry("IdInputScreen")
         }
-        val viewModel:SignUpViewModel = viewModel(parentEntry)
+        val viewModel: SignUpViewModel = viewModel(parentEntry)
         PasswordInputScreen(navController, viewModel)
     }
 }
