+++ app/src/main/java/org/sopt/at/presentation/history/HistoryScreen.kt
@@ -1,4 +1,4 @@
-package org.sopt.at.live
+package org.sopt.at.presentation.history
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
@@ -11,14 +11,14 @@ import androidx.navigation.NavController
 import org.sopt.at.utils.BottomNavigation
 
 @Composable
-fun LiveScreen(navController: NavController){
+fun HistoryScreen(navController: NavController){
     Box(
         modifier = Modifier
             .fillMaxSize()
             .background(color = Color.Black)
     ){
         Text(
-            text = "Live Screen"
+            text = "History Screen"
         )
         BottomNavigation(navController = navController)
     }
