+++ app/src/main/java/org/sopt/at/presentation/home/HomeScreen.kt
@@ -1,4 +1,4 @@
-package org.sopt.at.home
+package org.sopt.at.presentation.home
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
